requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.3
python-dotenv==1.0.0
PyPDF2==3.0.1
openpyxl==3.11.0
sqlparse==0.4.4
requests==2.31.0
